fix: Use more interceptions from edf api
With an iphone, on the first execution, our call to getBillsDocuments in
edf api returns an empty list of documents, which leads to an empty list
of bills, it may be a problem of order of calls to the api but the simplest
way to fix it now is to use interceptions of the json document. Edf itself
will handle the proper order of calls.